refactor(NumberInput): consolidate F0AmountCalculator into NumberInput
BREAKING CHANGE: Removes F0AmountCalculator component. Use NumberInput
with extraContent, inputWidth, and popover props instead.
