--- packages/react/src/experimental/Forms/Fields/F0AmountCalculator/__stories__/F0AmountCalculator.mdx
@@ -1,226 +1,0 @@
-import { Canvas, Meta, Controls, Unstyled } from "@storybook/addon-docs/blocks"
-import * as Stories from "./F0AmountCalculator.stories"
-import { DoDonts } from "@/lib/storybook-utils/do-donts"
-import { F0AmountCalculator } from "../F0AmountCalculator"
-import { I18nProvider, defaultTranslations } from "@/lib/providers/i18n"
-
-<Meta of={Stories} />
-
-# F0AmountCalculator
-
-A numeric input that supports two rendering modes: **inline** as a standard form field, or **popover** as a floating panel triggered by an icon button. Ideal for amount, percentage, or any numeric value that benefits from inline context (e.g. "of 300,00 €").
-
-## Positioning
-
-- Use `F0AmountCalculator` in bespoke product surfaces where you need **inline contextual amount text** (value + "of X") or **popover editing**.
-- In `F0Form`, prefer regular `number` / `percentage` fields for consistency, and use `helpText` for supportive text under the input.
-- Do not use `F0AmountCalculator` only to display a plain percentage input with no contextual right-side content.
-
-## Anatomy
-
-<Canvas of={Stories.Default} />
-
-<Controls of={Stories.Default} />
-
-### Mode
-
-<Canvas of={Stories.Inline} />
-
-<Unstyled>
-  <table className="mb-8 w-full dark:text-f1-foreground-inverse/80">
-    <thead>
-      <tr>
-        <th className="text-left">Mode</th>
-        <th className="text-left">Description</th>
-        <th className="text-left">When to use</th>
-      </tr>
-    </thead>
-    <tbody>
-      <tr>
-        <td>
-          <code>inline</code> (default)
-        </td>
-        <td>Renders a NumberInput directly in the layout</td>
-        <td>When the field is always visible in a form or table row</td>
-      </tr>
-      <tr>
-        <td>
-          <code>popover</code>
-        </td>
-        <td>
-          An outline icon button opens a floating panel containing the input
-        </td>
-        <td>
-          Dense surfaces where the input should be hidden until needed — table
-          cells, compact action rows
-        </td>
-      </tr>
-    </tbody>
-  </table>
-</Unstyled>
-
-### Popover
-
-<Canvas of={Stories.AsPopover} />
-
-Pass `popover={{}}` to activate popover mode with all defaults (Calculator icon, `bottom`/`start` placement, uncontrolled). Use `popover.icon` to swap the trigger icon, and `popover.open` + `popover.onOpenChange` for controlled open state.
-
-If you need an icon-only trigger while preserving accessibility, pass `hideLabel` and keep `label` defined. The label stays as the accessible name for the trigger button.
-
-<Canvas of={Stories.AsPopoverIconOnlyA11y} />
-
-<Canvas of={Stories.AsPopoverControlled} />
-
-In deferred commit mode (`popover.commitMode = "deferred"`), you can choose where the Apply action renders:
-
-- `popover.apply.layout = "block"` (default): dedicated action row below the input
-- `popover.apply.layout = "inline"`: Apply button on the same row as input/context
-
-<Canvas of={Stories.AsPopoverDeferredApply} />
-
-<Canvas of={Stories.AsPopoverDeferredApplyInline} />
-
-### States
-
-<Canvas of={Stories.States} />
-
-<Unstyled>
-  <table className="mb-8 w-full dark:text-f1-foreground-inverse/80">
-    <thead>
-      <tr>
-        <th className="text-left">State</th>
-        <th className="text-left">Description</th>
-        <th className="text-left">How to trigger</th>
-      </tr>
-    </thead>
-    <tbody>
-      <tr>
-        <td>
-          <code>default</code>
-        </td>
-        <td>Standard interactive state</td>
-        <td>No extra props</td>
-      </tr>
-      <tr>
-        <td>
-          <code>hint</code>
-        </td>
-        <td>Helper text below the input</td>
-        <td>
-          <code>hint="..."</code>
-        </td>
-      </tr>
-      <tr>
-        <td>
-          <code>error</code>
-        </td>
-        <td>Red border, error message below the input</td>
-        <td>
-          <code>error="..."</code> or <code>error={"{true}"}</code>
-        </td>
-      </tr>
-      <tr>
-        <td>
-          <code>warning</code>
-        </td>
-        <td>Warning message below the input</td>
-        <td>
-          <code>{'status={{ type: "warning", message: "..." }}'}</code>
-        </td>
-      </tr>
-      <tr>
-        <td>
-          <code>disabled</code>
-        </td>
-        <td>Inactive — does not respond to interaction</td>
-        <td>
-          <code>disabled</code>
-        </td>
-      </tr>
-    </tbody>
-  </table>
-</Unstyled>
-
-## Guidelines
-
-### Design best practices
-
-**When to use**
-
-- Use inline mode when the numeric field is always visible and is a primary data point in the form
-- Use popover mode on dense surfaces (table cells, compact rows) where showing the input inline would break the layout
-- Use `extraContent` to provide units or contextual values directly next to the input (e.g. `"of 300,00 €"`)
-- Use `inputWidth` to keep the input compact and prevent it from stretching across the available width
-- In `F0Form`, keep using `number` / `percentage` and put contextual explanation in `helpText`
-
-**When not to use**
-
-- For simple number inputs without context — use `NumberInput` directly
-- In `F0Form` when you only need below-input explanatory text — use field `helpText`
-- As a primary CTA button — the popover trigger is an icon button for revealing an input, not a general action trigger
-- For inputs that require rich formatting (currency symbols, thousands separators configured per locale) without passing `locale` — always pass a locale
-
-**Layout**
-
-- In inline mode with `extraContent`, the label and messages are automatically hoisted above and below the row — no extra wrapper needed
-- Keep `inputWidth` fixed when `extraContent` is present to prevent the row from jumping as the value changes
-- In popover mode, the panel has a minimum width of `220px` (`min-w-[220px]`) and adapts to content
-
-<DoDonts
-  do={{
-    description:
-      "Fix the input width when using extraContent to keep the row stable",
-    children: (
-      <I18nProvider translations={defaultTranslations}>
-        <F0AmountCalculator
-          label="Discount"
-          locale="en-US"
-          units="%"
-          inputWidth="160px"
-          extraContent="of 300,00 €"
-          onChange={() => {}}
-        />
-      </I18nProvider>
-    ),
-  }}
-  dont={{
-    description:
-      "Don't leave inputWidth unset when extraContent shifts the layout",
-    children: (
-      <I18nProvider translations={defaultTranslations}>
-        <F0AmountCalculator
-          label="Discount"
-          locale="en-US"
-          units="%"
-          extraContent="of 300,00 €"
-          onChange={() => {}}
-        />
-      </I18nProvider>
-    ),
-  }}
-/>
-
-### Content best practices
-
-- Keep `extraContent` short — a currency amount, a unit label, or a percentage base
-- Label text should follow the same rules as other form fields: sentence case, imperative where possible
-- Error messages should be specific: `"Value must be between 0 and 100"` not `"Invalid value"`
-
-### Behavior
-
-**Inline with extra content**
-
-When `extraContent` is provided, label, hint, and error are rendered outside the flex row. The inner `NumberInput` receives `hideLabel` and suppressed hint/error so the row height stays consistent.
-
-<Canvas of={Stories.InlineWithStep} />
-
-**Popover with custom icon**
-
-<Canvas of={Stories.AsPopoverWithCustomIcon} />
-
-## Accessibility
-
-- In popover mode the trigger button always requires a `label` prop — it is used as the accessible name and tooltip even when the icon is self-explanatory
-- The Calculator icon is the default trigger; override with `popover.icon` when the context suggests a different affordance
-- Error and warning messages are rendered below the input and announced via the standard `InputMessages` component
-- `disabled` uses native disabled semantics — the input is removed from tab order
