--- packages/react/src/experimental/Forms/Fields/F0AmountCalculator/index.tsx
@@ -1,2 +1,0 @@
-export * from "./F0AmountCalculator"
-export * from "./types"
