+++ packages/react/src/experimental/Forms/Fields/exports.ts
@@ -1,7 +1,6 @@
 export { F0Select as Select } from "../../../components/F0Select"
 export * from "../../../components/F0Select/types"
 export * from "./F1SearchBox"
-export * from "./F0AmountCalculator"
 export * from "./Input"
 export * from "./NumberInput"
 export * from "./Switch"
